requirements.txt
absl-py==2.1.0
accelerate==0.33.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.38.1
bleach==6.1.0
blessed==1.20.0
boto3==1.35.7
botocore==1.35.7
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.25.0
coloredlogs==15.0.1
colorlog==6.8.2
comm==0.2.2
contourpy==1.3.0
controlnet-aux==0.0.9
croniter==1.3.15
cycler==0.12.1
dataclasses-json==0.6.7
dateutils==0.6.12
dearpygui==1.11.1
debugpy==1.8.5
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
Deprecated==1.2.14
diffusers==0.23.0
docker-pycreds==0.4.0
docstring_parser==0.16
editor==1.6.6
einops==0.8.0
embreex==2.17.7.post5
envlight @ git+https://github.com/ashawkey/envlight.git@05b5851e854429d72ecaf5b206ed64ce55fae677
exceptiongroup==1.2.2
executing==2.0.1
fastapi==0.116.0
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.18.0
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2025.5.1
ftfy==6.2.3
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
GPUtil==1.4.0
gradio==5.36.2
gradio_client==1.10.4
groovy==0.1.2
grpcio==1.66.0
h11==0.14.0
hf-xet==1.1.5
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.33.2
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.4.4
inquirer==3.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.31
jaxlib==0.4.31
jaxtyping==0.2.33
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiui==0.2.11
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
libigl==2.5.1
lightning==2.0.0
lightning-cloud==0.5.70
lightning-utilities==0.11.6
lit==15.0.7
llvmlite==0.43.0
loguru==0.7.2
lxml==5.3.0
manifold3d==2.5.1
mapbox_earcut==1.0.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.14
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
narwhals==1.46.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nerfacc @ git+https://github.com/KAIR-BAIR/nerfacc.git@d84cdf3afd7dcfc42150e0f0506db58a5ce62812
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nodeenv==1.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.3
nvdiffrast @ git+https://github.com/NVlabs/nvdiffrast.git@c5caf7bdb8a2448acc491a9faa47753972edd380
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
objaverse==0.1.7
objprint==0.2.3
omegaconf==2.3.0
onnxruntime==1.19.2
open_clip_torch==2.30.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==25.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.2
plotly==6.2.0
pooch==1.8.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pycollada==0.8
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
pygltflib==1.16.2
Pygments==2.18.0
PyJWT==2.9.0
pyliblzfse==0.4.1
PyMatting==1.1.12
PyMCubes==0.1.6
pymeshlab==2023.12.post1
pyparsing==3.1.4
pyre-extensions==0.0.29
pysdf==0.1.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.20
pytorch-lightning==2.3.3
pytz==2024.1
pyvista==0.45.2
PyWavefront==1.3.3
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.5.2
QtPy==2.4.1
readchar==4.2.0
referencing==0.35.1
regex==2024.7.24
rembg==2.0.59
requests==2.32.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.0
rpds-py==0.20.0
Rtree==1.3.0
ruff==0.12.2
runs==1.2.2
s3transfer==0.10.2
safehttpx==0.1.6
safetensors==0.4.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
scooby==0.10.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.13.0
setproctitle==1.3.3
shapely==2.0.6
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.2
starsessions==1.3.0
svg.path==6.3
sympy==1.12
taming-transformers-rom1504==0.0.6
tensorboard==2.17.1
tensorboard-data-server==0.7.2
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2024.8.24
timm==0.6.7
tinycss2==1.3.0
tinycudann @ git+https://github.com/NVlabs/tiny-cuda-nn/@c91138bcd4c6877c8d5e60e483c0581aafc70cce#subdirectory=bindings/torch
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.12.0
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchmetrics==1.4.1
torchvision==0.15.2+cu118
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.28.1
trimesh==4.4.7
triton==2.0.0
typeguard==2.13.3
typer==0.12.5
types-python-dateutil==2.9.0.20240821
typing-inspect==0.9.0
typing_extensions==4.14.1
tyro==0.8.10
tzdata==2024.1
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.30.6
varname==0.13.3
vhacdx==0.0.8.post1
viser==0.2.3
vtk==9.4.2
wandb==0.17.7
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
xatlas==0.0.9
xformers==0.0.20
xmod==1.8.1
xxhash==3.5.0
yarl==1.9.4
yourdfpy==0.0.56
zipp==3.20.1
